worker/requirements.txt
cotyledon
kafka-python
influxdb-client